+++ ImgReconstruction/Core/CFFTBlurMeasurer.hpp
@@ -13,14 +13,14 @@
 class CFFTBlurMeasurer : public IBlurMeasurer
 {
 public:
-	CFFTBlurMeasurer(double radiusRatio) : _radiusRatio(radiusRatio) {};
+	CFFTBlurMeasurer(float radiusRatio) : _radiusRatio(radiusRatio) {};
 
-	virtual double Measure(const CImage& img) const
+	virtual float Measure(const CImage& img) const
 	{
 		CImage fft = img.GetFFTImage();
 		img.CopyMetadataTo(fft);
 
-		cv::Size submatrixSize = cv::Size(ceil(img.cols * _radiusRatio), ceil(img.rows * _radiusRatio));
+		cv::Size submatrixSize = cv::Size((int)ceil(img.cols * _radiusRatio), (int)ceil(img.rows * _radiusRatio));
 		cv::Point submatrixOrigin = cv::Point((img.cols - submatrixSize.width) / 2, (img.rows - submatrixSize.height) / 2);
 		cv::Rect submatrixRect = cv::Rect(submatrixOrigin, submatrixSize);
 
@@ -30,9 +30,9 @@ public:
 		CImage roi = imageCopy(submatrixRect);
 		roi.setTo(0);
 
-		return -cv::sum(imageCopy)[0];
+		return (float)-cv::sum(imageCopy)[0];
 	}
 
 private:
-	double _radiusRatio;
+	float _radiusRatio;
 };
