+++ ImgReconstruction/Core/CInterpolationKernel.hpp
@@ -7,12 +7,12 @@ class CInterpolationKernel: public IInterpolationKernel
 public:
 	CInterpolationKernel(int a) : _a(a) {};
 
-	virtual double operator()(double x) const = 0;
+	virtual float operator()(float x) const = 0;
 
-	virtual std::vector<double> Coeffs(double shift) const
+	virtual std::vector<float> Coeffs(float shift) const
 	{
-		std::vector<double> coeffs;
-		double floorx = floor(shift);
+		std::vector<float> coeffs;
+		float floorx = floorf(shift);
 		int a = this->A();
 		for (int i = 0; i <= 2 * a - 1; i++) {
 			coeffs.push_back((*this)(shift - (i + floorx - a + 1)));
